abstract out intializeMenus
