+++ src/notebook/menuDivider.js
@@ -1,0 +1,4 @@
+export const menuDivider = {
+  label: null,
+  action: null,
+};
