+++ src/notebook/notebook.js
@@ -38,11 +38,8 @@ import { runCellAndSelectBelow } from './runCellAndSelectBelow';
 import { saveNotebook } from './saveNotebook';
 import { loadNotebook } from './loadNotebook';
 import { promptForNotebook } from './promptForNotebook';
-import { uploadFile } from './uploadFile';
 import { toggleInput } from './toggleInput';
 import { toggleOutput } from './toggleOutput';
-import { toggleAllInputs } from './toggleAllInputs';
-import { toggleAllOutputs } from './toggleAllOutputs';
 import { editName } from './editName';
 import { saveName } from './saveName';
 import { toggleSidebar } from './toggleSidebar';
@@ -50,28 +47,17 @@ import { toggleSidebar } from './toggleSidebar';
 // hence no {} curly braces
 import selectPreviousCell from './selectPreviousCell';
 import displayKeyboardShortcuts from './displayKeyboardShortcuts';
-import displayDocumentation from './displayDocumentation';
-import displayFAQ from './displayFAQ';
 import executeCommand from './executeCommand';
-import displayAbout from './displayAbout';
-import shutdown from './shutdown';
-import showHelp from './showHelp';
 import createNotebook from './createNotebook';
-import duplicateNotebook from './duplicateNotebook';
 import openNotebook from './openNotebook';
-import exportNotebook from './exportNotebook';
-import goToH2OUrl from './goToH2OUrl';
-import goToUrl from './goToUrl';
 import executeAllCells from './executeAllCells';
 import runAllCells from './runAllCells';
-import continueRunningAllCells from './continueRunningAllCells';
 import stopRunningAll from './stopRunningAll';
 import clearCell from './clearCell';
-import clearAllCells from './clearAllCells';
 import notImplemented from './notImplemented';
 import createMenu from './createMenu';
-import createMenuHeader from './createMenuHeader';
 import createMenuItem from './createMenuItem';
+import initializeMenus from './initializeMenus';
 
 import { requestModelBuilders } from '../h2oProxy/requestModelBuilders';
 import { getObjectExistsRequest } from '../h2oProxy/getObjectExistsRequest';
@@ -80,6 +66,9 @@ import { flowStatus } from '../flowStatus';
 import { flowSidebar } from '../flowSidebar';
 import { flowPreludeFunction } from '../flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
+
+// constants
+import { menuDivider } from './menuDivider';
 
 export function notebook() {
   const lodash = window._;
@@ -101,8 +90,8 @@ export function notebook() {
     _.selectedCellIndex = -1;
     _.clipboardCell = null;
     _.lastDeletedCell = null;
-    const _areInputsHidden = Flow.Dataflow.signal(false);
-    const _areOutputsHidden = Flow.Dataflow.signal(false);
+    _.areInputsHidden = Flow.Dataflow.signal(false);
+    _.areOutputsHidden = Flow.Dataflow.signal(false);
     _.isSidebarHidden = Flow.Dataflow.signal(false);
     _.isRunningAll = Flow.Dataflow.signal(false);
     _.runningCaption = Flow.Dataflow.signal('Running');
@@ -119,10 +108,6 @@ export function notebook() {
     //
     // Top menu bar
     //
-    const menuDivider = {
-      label: null,
-      action: null,
-    };
     const _menus = Flow.Dataflow.signal(null);
     menuCell = [
       createMenuItem('Run Cell', runCell.bind(this, _), [
@@ -171,96 +156,11 @@ export function notebook() {
     if (_.onSparklingWater) {
       menuCell = __slice.call(menuCell).concat(__slice.call(menuCellSW));
     }
-    const initializeMenus = builder => {
-      const modelMenuItems = lodash.map(builder, builder => createMenuItem(`${builder.algo_full_name}...`, executeCommand(_, `buildModel ${flowPrelude.stringify(builder.algo)}`))).concat([
-        menuDivider,
-        createMenuItem('List All Models', executeCommand(_, 'getModels')),
-        createMenuItem('List Grid Search Results', executeCommand(_, 'getGrids')),
-        createMenuItem('Import Model...', executeCommand(_, 'importModel')),
-        createMenuItem('Export Model...', executeCommand(_, 'exportModel')),
-      ]);
-      return [
-        createMenu('Flow', [
-          createMenuItem('New Flow', createNotebook.bind(this, _)),
-          createMenuItem('Open Flow...', promptForNotebook.bind(this, _)),
-          createMenuItem('Save Flow', saveNotebook.bind(this, _), ['s']),
-          createMenuItem('Make a Copy...', duplicateNotebook.bind(this, _)),
-          menuDivider,
-          createMenuItem('Run All Cells', runAllCells.bind(this, _)),
-          createMenuItem('Run All Cells Below', continueRunningAllCells.bind(this, _)),
-          menuDivider,
-          createMenuItem('Toggle All Cell Inputs', toggleAllInputs.bind(this, _, _areInputsHidden)),
-          createMenuItem('Toggle All Cell Outputs', toggleAllOutputs.bind(this, _areOutputsHidden)),
-          createMenuItem('Clear All Cell Outputs', clearAllCells.bind(this, _)),
-          menuDivider,
-          createMenuItem('Download this Flow...', exportNotebook.bind(this, _)),
-        ]),
-        createMenu('Cell', menuCell),
-        createMenu('Data', [
-          createMenuItem('Import Files...', executeCommand(_, 'importFiles')),
-          createMenuItem('Upload File...', uploadFile.bind(this, _)),
-          createMenuItem('Split Frame...', executeCommand(_, 'splitFrame')),
-          createMenuItem('Merge Frames...', executeCommand(_, 'mergeFrames')),
-          menuDivider,
-          createMenuItem('List All Frames', executeCommand(_, 'getFrames')),
-          menuDivider,
-          createMenuItem('Impute...', executeCommand(_, 'imputeColumn')),
-          // TODO Quantiles
-          // TODO Interaction
-        ]),
-        createMenu('Model', modelMenuItems),
-        createMenu('Score', [
-          createMenuItem('Predict...', executeCommand(_, 'predict')),
-          createMenuItem('Partial Dependence Plots...', executeCommand(_, 'buildPartialDependence')),
-          menuDivider,
-          createMenuItem('List All Predictions', executeCommand(_, 'getPredictions')),
-          // TODO Confusion Matrix
-          // TODO AUC
-          // TODO Hit Ratio
-          // TODO PCA Score
-          // TODO Gains/Lift Table
-          // TODO Multi-model Scoring
-        ]),
-        createMenu('Admin', [
-          createMenuItem('Jobs', executeCommand(_, 'getJobs')),
-          createMenuItem('Cluster Status', executeCommand(_, 'getCloud')),
-          createMenuItem('Water Meter (CPU meter)', goToH2OUrl('perfbar.html')),
-          menuDivider,
-          createMenuHeader('Inspect Log'),
-          createMenuItem('View Log', executeCommand(_, 'getLogFile')),
-          createMenuItem('Download Logs', goToH2OUrl('3/Logs/download')),
-          menuDivider,
-          createMenuHeader('Advanced'),
-          createMenuItem('Create Synthetic Frame...', executeCommand(_, 'createFrame')),
-          createMenuItem('Stack Trace', executeCommand(_, 'getStackTrace')),
-          createMenuItem('Network Test', executeCommand(_, 'testNetwork')),
-          // TODO Cluster I/O
-          createMenuItem('Profiler', executeCommand(_, 'getProfile depth: 10')),
-          createMenuItem('Timeline', executeCommand(_, 'getTimeline')),
-          // TODO UDP Drop Test
-          // TODO Task Status
-          createMenuItem('Shut Down', shutdown.bind(this, _)),
-        ]),
-        createMenu('Help', [
-          // TODO createMenuItem('Tour', startTour, true),
-          createMenuItem('Assist Me', executeCommand(_, 'assist')),
-          menuDivider,
-          createMenuItem('Contents', showHelp.bind(this, _)),
-          createMenuItem('Keyboard Shortcuts', displayKeyboardShortcuts, ['h']),
-          menuDivider,
-          createMenuItem('Documentation', displayDocumentation),
-          createMenuItem('FAQ', displayFAQ),
-          createMenuItem('H2O.ai', goToUrl('http://h2o.ai/')),
-          createMenuItem('H2O on Github', goToUrl('https://github.com/h2oai/h2o-3')),
-          createMenuItem('Report an issue', goToUrl('http://jira.h2o.ai')),
-          createMenuItem('Forum / Ask a question', goToUrl('https://groups.google.com/d/forum/h2ostream')),
-          menuDivider,
-          // TODO Tutorial Flows
-          createMenuItem('About', displayAbout),
-        ]),
-      ];
-    };
-    const setupMenus = () => requestModelBuilders(_, (error, builders) => _menus(initializeMenus(error ? [] : builders)));
+    const setupMenus = () => requestModelBuilders(_, (error, builders) => _menus(initializeMenus(
+      _,
+      menuCell,
+      error ? [] : builders
+    )));
     const createTool = (icon, label, action, isDisabled) => {
       if (isDisabled == null) {
         isDisabled = false;
@@ -587,8 +487,8 @@ export function notebook() {
       status: _status,
       toolbar: _toolbar,
       cells: _.cells,
-      areInputsHidden: _areInputsHidden,
-      areOutputsHidden: _areOutputsHidden,
+      areInputsHidden: _.areInputsHidden,
+      areOutputsHidden: _.areOutputsHidden,
       isSidebarHidden: _.isSidebarHidden,
       isRunningAll: _.isRunningAll,
       runningCaption: _.runningCaption,
